+++ doc/change-log.md
@@ -34,6 +34,8 @@
 - `StrmStatus` 增加 `sent` 属性
 - 为了避免前端 js 解释时不丢失精度，`TermConnState` 的 `connAt`、`lastActive` 改为 String 类型
 - `GnssOpenStrmResult`, `OpenStrmResult` 增加 `channel` 属性
+- `StrmMediaNotif` 增加 `aBitrate`, `frameRate` 属性
+- 新增 `PagedQuery`, `QryStrmLogReq` 类
 
 ## 不兼容变更
 - `AvUploadReq` 增加 `fileSz` 必填属性